requirements.txt
discord.py[voice]
bs4
requests
youtube_dl
speedtest-cli